fix(SNegato2DCamera): fix conversion warnings

+++ Bundles/visu/visuOgreAdaptor/src/visuOgreAdaptor/SNegato2DCamera.cpp
@@ -291,25 +291,27 @@ void SNegato2DCamera::resetCamera()
         case Orientation::X_AXIS:
             camNode->rotate(::Ogre::Vector3::UNIT_Y, ::Ogre::Degree(-90.f));
             camNode->rotate(::Ogre::Vector3::UNIT_Z, ::Ogre::Degree(-90.f));
-            camera->setOrthoWindowHeight(static_cast< ::Ogre::Real >(size[2] * spacing[2]));
+            camera->setOrthoWindowHeight(static_cast< ::Ogre::Real >(static_cast<double>(size[2]) * spacing[2]));
             break;
         case Orientation::Y_AXIS:
             camNode->rotate(::Ogre::Vector3::UNIT_X, ::Ogre::Degree(90.f));
-            camera->setOrthoWindowHeight(static_cast< ::Ogre::Real >(size[2] * spacing[2]));
+            camera->setOrthoWindowHeight(static_cast< ::Ogre::Real >(static_cast<double>(size[2]) * spacing[2]));
             break;
         case Orientation::Z_AXIS:
             camNode->rotate(::Ogre::Vector3::UNIT_Z, ::Ogre::Degree(180.f));
             camNode->rotate(::Ogre::Vector3::UNIT_Y, ::Ogre::Degree(180.f));
-            camera->setOrthoWindowHeight(static_cast< ::Ogre::Real >(size[1] * spacing[1]));
+            camera->setOrthoWindowHeight(static_cast< ::Ogre::Real >(static_cast<double>(size[1]) * spacing[1]));
             break;
     }
 
     const size_t orientation = static_cast<size_t>(m_currentNegatoOrientation);
-    ::Ogre::Vector3 camPos( static_cast< ::Ogre::Real >(origin[0] + size[0] * spacing[0] * 0.5),
-                            static_cast< ::Ogre::Real >(origin[1] + size[1] * spacing[1] * 0.5),
-                            static_cast< ::Ogre::Real >(origin[2] + size[2] * spacing[2] * 0.5));
+    ::Ogre::Vector3 camPos( static_cast< ::Ogre::Real >(origin[0] + static_cast<double>(size[0]) * spacing[0] * 0.5),
+                            static_cast< ::Ogre::Real >(origin[1] + static_cast<double>(size[1]) * spacing[1] * 0.5),
+                            static_cast< ::Ogre::Real >(origin[2] + static_cast<double>(size[2]) * spacing[2] * 0.5));
 
-    camPos[orientation] = static_cast< ::Ogre::Real >(origin[orientation] - size[orientation] * spacing[orientation]);
+    camPos[orientation] =
+        static_cast< ::Ogre::Real >(origin[orientation] - static_cast<double>(size[orientation]) *
+                                    spacing[orientation]);
     camNode->setPosition(camPos);
 
     const auto worldBoundingBox = layer->computeWorldBoundingBox();
